+++ src/core.js
@@ -6,7 +6,7 @@
 */
 
 const LX = {
-    version: "0.6.8",
+    version: "0.6.9",
     ready: false,
     components: [], // Specific pre-build components
     signals: {}, // Events and triggers