+++ changelog.md
@@ -13,6 +13,7 @@ Widgets:
 - Fixed disabled NumberInput slider.
 - TabSections icons can be modified using `iconClass` and `svgClass`.
 - TabSection's `onCreate` now receives also the tab content HTML Container (2nd parameter).
+- Fixed some Delete edge cases (key and ContextMenu) of Tree Widget.
 
 ## 0.6.8 (master)